backend/requirements.txt
amqp==5.3.1
asgiref==3.8.1
beautifulsoup4==4.13.4
billiard==4.2.1
Brotli==1.1.0
bs4==0.0.2
celery==5.5.3
cffi==1.17.1
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
confluent-kafka==2.10.1
cssselect2==0.8.0
dj-database-url==3.0.1
Django==5.2
django-cors-headers==4.3.1
django-filter==25.1
django-rest-framework==0.1.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
fonttools==4.57.0
gunicorn==23.0.0
kafka-python==2.2.12
kombu==5.5.4
packaging==25.0
pillow==11.2.1
prompt_toolkit==3.0.51
psycopg2-binary==2.9.9
pycparser==2.22
pydyf==0.11.0
PyJWT==2.10.1
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
redis==6.2.0
six==1.17.0
soupsieve==2.6
sqlparse==0.5.3
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.13.2
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.13
weasyprint==65.1
webencodings==0.5.1
zopfli==0.2.3.post1